# Reduce kernel calls for water sampling