Speeds up targeted insertion by about 20%. These are the simpler changes.


## Kernel Performance
Ran following test: `pytest tests/test_cuda_targeted_insertion_mover.py -k 'test_moves_with_complex_and_ligand[2023-float32-0.0001-0.002-12500-12500-2.0]'`
Cuda Arch 8.6
GPU: A4000

### Changes

* `void timemachine::k_set_sampled_weight` previously took 20k ns per run, split into two kernels each kernel is only 4k ns each thanks to additional parallelism.
* `void timemachine::k_flag_mols_inner_outer` is now only run once, reducing 1.4% kernel run time

### Old
```
CUDA API Statistics:

 Time (%)  Total Time (ns)  Num Calls  Avg (ns)   Med (ns)   Min (ns)   Max (ns)    StdDev (ns)            Name           
 --------  ---------------  ---------  ---------  ---------  --------  -----------  -----------  -------------------------
     51.0    1,793,051,451    449,433    3,989.6    3,734.0     3,043    1,218,286      2,511.8  cudaLaunchKernel         
     22.7      798,663,183     25,013   31,929.9    1,787.0     1,467      347,431     65,191.8  cudaEventSynchronize     
      8.4      296,085,609     66,162    4,475.2    3,930.0     2,247      487,328      7,847.4  cudaMemcpyAsync          
      6.4      224,447,919      2,017  111,278.1    7,092.0       131  177,337,124  3,948,571.0  cudaMalloc               
      2.7       95,882,718     18,827    5,092.8    4,169.0     2,030      160,571      4,398.3  cudaMemsetAsync          
      2.6       92,579,194      1,175   78,790.8   48,487.0     5,311    1,333,294    119,639.1  cudaMemcpy               
      2.4       83,363,817    113,346      735.5      591.0       429       20,854        617.7  cudaEventRecord          
      2.0       70,461,715     88,420      796.9      650.0       464       19,230        667.7  cudaStreamWaitEvent      
      1.2       42,871,077      2,037   21,046.2    5,717.0       312   11,492,198    258,040.3  cudaFree                 
      0.1        5,270,871        392   13,446.1    6,890.5     1,736      420,042     30,833.5  cudaStreamSynchronize    
      0.1        4,177,442          5  835,488.4  166,974.0    97,309    3,341,631  1,406,091.4  cudaDeviceSynchronize    
      0.1        3,779,216         11  343,565.1   11,075.0     6,748      951,056    462,976.2  cudaHostAlloc            
      0.1        2,731,243         11  248,294.8   22,248.0     5,687    1,117,100    366,668.7  cudaFreeHost             
      0.0          319,140         37    8,625.4    2,920.0     2,582      146,537     23,715.8  cudaStreamCreateWithFlags
      0.0          227,646         37    6,152.6    3,789.0     2,693       26,548      5,000.8  cudaStreamDestroy        
      0.0          206,135         22    9,369.8    5,272.0     2,343       30,115      7,624.9  cudaMemset               
      0.0          188,107        756      248.8      225.5       134        3,329        145.5  cuGetProcAddress         
      0.0           68,142         47    1,449.8      865.0       419        4,819      1,092.7  cudaEventDestroy         
      0.0           62,089         47    1,321.0      851.0       505        3,075        857.9  cudaEventCreateWithFlags 
      0.0            5,978          2    2,989.0    2,989.0     2,888        3,090        142.8  cuInit                   
      0.0              442          2      221.0      221.0       176          266         63.6  cuModuleGetLoadingMode   

[5/7] Executing 'gpukernsum' stats report

CUDA Kernel Statistics:

 Time (%)  Total Time (ns)  Instances   Avg (ns)     Med (ns)    Min (ns)   Max (ns)   StdDev (ns)                                                  Name                                                
 --------  ---------------  ---------  -----------  -----------  ---------  ---------  -----------  ----------------------------------------------------------------------------------------------------
     32.0      875,510,612     10,809     80,998.3     26,112.0     15,168    212,960     82,075.4  void timemachine::k_nonbonded_unified<float, (int)256, (bool)0, (bool)1, (bool)0>(int, int, const u…
      9.4      256,673,574     12,500     20,533.9     20,512.0     19,424     21,504        262.2  void timemachine::k_set_sampled_weight<float, (int)512>(int, int, const int *, const int *, const i…
      4.2      114,834,551     25,000      4,593.4      4,576.0      4,385     12,993        167.5  void timemachine::k_atom_by_atom_energies<float>(int, int, const int *, const double *, const doubl…
      3.5       96,536,551      1,200     80,447.1     25,728.0     21,952    217,856     78,319.0  void timemachine::k_nonbonded_unified<float, (int)256, (bool)1, (bool)0, (bool)0>(int, int, const u…
      2.5       69,135,838        559    123,677.7    124,577.0    107,008    143,936      8,428.7  void timemachine::k_find_blocks_with_ixns<float, (bool)1>(int, int, int, const unsigned int *, cons…
      2.4       66,868,869      1,019     65,622.1     57,183.0     41,792    144,704     19,836.0  void timemachine::k_find_blocks_with_ixns<float, (bool)0>(int, int, int, const unsigned int *, cons…
      2.4       66,416,429     12,426      5,345.0      4,383.0      2,432     18,592      1,904.5  void timemachine::k_check_rebuild_coords_and_box_gather<float>(int, const unsigned int *, const dou…
      2.2       61,331,720     25,001      2,453.2      2,528.0      2,207     13,504        233.9  void cub::DeviceReduceSingleTileKernel<cub::DeviceReducePolicy<float, float, int, cub::Max>::Policy…
      2.2       60,777,227      4,171     14,571.4     13,248.0     10,400     37,472      4,142.6  void timemachine::k_nonbonded_precomputed<float>(int, const double *, const double *, const double …
      2.2       60,602,612        504    120,243.3     38,512.0     29,216    297,504    121,432.2  void timemachine::k_nonbonded_unified<float, (int)256, (bool)1, (bool)1, (bool)0>(int, int, const u…
      2.2       60,212,386     25,001      2,408.4      2,527.0      2,175     13,216        228.7  void cub::DeviceReduceSingleTileKernel<cub::DeviceReducePolicy<float, float, int, cub::Sum>::Policy…
      1.8       48,095,748      4,171     11,531.0     11,488.0      5,632     24,352      1,415.0  void timemachine::k_nonbonded_pair_list<float, (bool)1>(int, const double *, const double *, const …
      1.7       47,272,312     12,500      3,781.8      3,776.0      3,711     13,152        289.8  void timemachine::k_adjust_weights<float, (bool)1>(int, int, int, const int *, const int *, const T…
      1.7       47,269,592     25,001      1,890.7      1,888.0      1,695     12,672        196.0  void gen_sequenced<curandStateXORWOW, float, int, &curand_uniform_noargs<curandStateXORWOW>, rng_co…
      1.7       47,256,556     12,500      3,780.5      3,776.0      3,392     12,736        241.6  void timemachine::k_adjust_weights<float, (bool)0>(int, int, int, const int *, const int *, const T…
      1.7       46,195,008     12,513      3,691.8      3,680.0      2,016     14,816        648.9  void timemachine::k_gather_coords_and_params<double, (int)3, (int)4>(int, const unsigned int *, con…
      1.6       44,752,179     25,001      1,790.0      1,761.0      1,472     12,960        190.1  void timemachine::k_exp_sub_max<float>(int, const T1 *, const T1 *, T1 *)                           
      1.5       41,902,140     12,500      3,352.2      3,328.0      3,199     13,408        111.9  void timemachine::k_rotate_and_translate_mols<float, (bool)0>(int, const double *, const double *, …
      1.5       40,107,947     12,500      3,208.6      3,200.0      2,816      3,680         60.6  void cub::DeviceSelectSweepKernel<cub::DispatchSelectIf<int *, int *, int *, int *, cub::NullType, …
      1.4       39,429,407     12,500      3,154.4      3,136.0      3,071     12,832        135.1  void timemachine::k_flag_mols_inner_outer<float>(int, const int *, const int *, const T1 *, T1, con…
      1.3       35,171,424      4,171      8,432.4      8,256.0      3,776     18,784      1,798.7  void timemachine::k_periodic_torsion<float, (int)3>(int, const double *, const double *, const int …
      1.2       34,214,426     11,313      3,024.3      3,136.0      1,632     14,656        725.4  void timemachine::k_scatter_accum<unsigned long long, (int)3>(int, const unsigned int *, const T1 *…
      1.2       33,155,558     12,500      2,652.4      2,784.0      2,431     12,992        220.7  void cub::DeviceReduceSingleTileKernel<cub::DeviceReducePolicy<cub::KeyValuePair<int, float>, cub::…
      1.1       30,036,309     12,500      2,402.9      2,400.0      2,271      8,448        101.8  void timemachine::k_generate_translations_within_or_outside_a_sphere<float>(int, const double *, co…
      1.0       28,030,023     12,500      2,242.4      2,240.0      2,176      2,656         49.1  timemachine::k_setup_sample_atoms(int, const int *, const int *, const int *, int *, int *)         
      1.0       27,924,209      6,248      4,469.3      3,871.0      2,304     20,960      1,595.4  void timemachine::k_accumulate_energy<(unsigned int)512>(int, const __int128 *, __int128 *)         
      1.0       27,289,816     12,500      2,183.2      2,144.0      1,920     12,896        463.2  void timemachine::k_setup_destination_weights_for_targeted<float>(int, const int *, const int *, co…
      1.0       27,153,906     12,850      2,113.1      1,920.0      1,823     18,305      1,063.0  void gen_sequenced<curandStateXORWOW, float2, normal_args_st, &curand_normal_scaled2<curandStateXOR…
      0.9       25,577,974     12,500      2,046.2      2,048.0      1,983      2,496         49.5  void timemachine::k_separate_weights_for_targeted<float>(int, const int *, const int *, const int *…
      0.9       25,531,796      4,171      6,121.3      5,665.0      3,392     17,504      1,432.4  void timemachine::k_harmonic_bond<float>(int, const double *, const double *, const int *, unsigned…
      0.9       24,962,028     12,500      1,997.0      1,984.0      1,600     12,768        108.4  void timemachine::k_decide_targeted_move<float>(int, const T1 *, const int *, int *)                
      0.9       24,643,923     12,499      1,971.7      1,952.0      1,919     10,303         97.5  void timemachine::k_store_accepted_log_probability_targeted<float>(int, const int *, const T1 *, T1…
      0.9       24,635,810     12,500      1,970.9      1,952.0      1,919      7,136         67.3  void timemachine::k_attempt_exchange_move_targeted<float>(int, const int *, const T1 *, T1, const T…
      0.9       24,589,002     12,500      1,967.1      1,952.0      1,919      2,400         48.9  timemachine::k_adjust_sample_idx(const int *, const int *, const int *, int *)                      
      0.9       23,300,062     12,500      1,864.0      1,856.0      1,823      2,273         53.5  void timemachine::k_setup_gumbel_max_trick<float>(int, const T1 *, T1 *)                            
      0.8       23,027,997     12,500      1,842.2      1,824.0      1,791      2,272         58.3  void timemachine::k_copy_kv_key<float>(int, const cub::KeyValuePair<int, T1> *, int *)              
      0.8       20,998,507     12,500      1,679.9      1,664.0      1,631     12,576        193.1  void cub::DeviceCompactInitKernel<cub::ScanTileState<int, (bool)1>, int *>(T1, int, T2)             
      0.7       19,885,371      3,500      5,681.5      5,697.0      4,992     13,792        435.9  void timemachine::k_update_forward_baoab<float, (int)3>(int, T1, const unsigned int *, const T1 *, …
      0.7       19,516,722      2,603      7,497.8      6,560.0      5,344     19,136      1,960.6  void timemachine::k_harmonic_angle<float, (int)3>(int, const double *, const double *, const int *,…
      0.5       14,895,783      2,597      5,735.8      5,664.0      3,872     16,064        865.1  void timemachine::k_find_block_bounds<float>(int, int, const unsigned int *, const double *, const …
      0.4        9,628,888      1,568      6,140.9      5,888.0      3,200     11,776      1,780.3  void timemachine::k_harmonic_angle_stable<float>(int, const double *, const double *, const int *, …
      0.3        7,024,776      1,578      4,451.7      4,576.0      2,240     15,648      1,375.8  timemachine::k_compact_trim_atoms(int, int, unsigned int *, unsigned int *, int *, unsigned int *)  
      0.2        4,358,114          1  4,358,114.0  4,358,114.0  4,358,114  4,358,114          0.0  void timemachine::k_compute_nonbonded_target_atom_energies<float, (int)64>(int, int, const int *, c…
      0.1        3,409,176        260     13,112.2     13,136.0      9,984     17,856      1,275.7  void cub::DeviceRadixSortOnesweepKernel<cub::DeviceRadixSortPolicy<unsigned int, unsigned int, int>…
      0.1        1,387,907         66     21,028.9     20,688.0     17,216     30,688      2,730.3  void cub::DeviceRadixSortSingleTileKernel<cub::DeviceRadixSortPolicy<unsigned int, unsigned int, in…
      0.0        1,293,277        200      6,466.4      6,304.0      5,920      7,744        435.8  void timemachine::k_find_group_centroids<float>(int, const double *, const int *, const int *, unsi…
      0.0        1,236,728        200      6,183.6      6,368.0      4,480      7,488        762.6  void timemachine::k_decide_move<float>(int, bool, int, double, double, const T1 *, T1 *, double *, …
      0.0          955,106        200      4,775.5      4,640.0      4,512      5,472        275.1  void timemachine::k_rescale_positions<float>(int, double *, const T1 *, const double *, double *, c…
      0.0          720,672        131      5,501.3      5,344.0      2,880     13,920      1,693.3  timemachine::k_coords_to_kv_gather(int, const unsigned int *, const double *, const double *, const…
      0.0          698,593          5    139,718.6    155,424.0     94,272    156,832     26,819.0  void generate_seed_pseudo<rng_config<curandStateXORWOW, (curandOrdering)101>>(unsigned long long, u…
      0.0          482,080        200      2,410.4      2,336.0      2,271      2,912        137.4  void timemachine::k_setup_barostat_move<float>(bool, const T1 *, double *, T1 *, double *, T1 *)    
      0.0          294,948         65      4,537.7      4,384.0      3,071      6,496        570.5  void cub::DeviceRadixSortHistogramKernel<cub::DeviceRadixSortPolicy<unsigned int, unsigned int, int…
      0.0          164,223         65      2,526.5      2,336.0      1,920      4,000        490.9  void cub::DeviceRadixSortExclusiveSumKernel<cub::DeviceRadixSortPolicy<unsigned int, unsigned int, …
      0.0          161,792          1    161,792.0    161,792.0    161,792    161,792          0.0  void timemachine::k_accumulate_atom_energies_to_per_mol_energies<float, (int)256>(int, const int *,…
      0.0           98,369         48      2,049.4      1,936.5      1,856      2,592        207.3  timemachine::k_arange(int, unsigned int *, unsigned int)                                            
      0.0           23,360         12      1,946.7      1,920.0      1,888      2,112         74.5  void timemachine::k_initialize_array<unsigned int>(int, T1 *, T1)                                   
      0.0            8,512          1      8,512.0      8,512.0      8,512      8,512          0.0  void timemachine::k_compute_centroid_of_atoms<float>(int, const int *, const double *, T1 *)        
      0.0            2,368          1      2,368.0      2,368.0      2,368      2,368          0.0  timemachine::k_initialize_curand_states(int, int, curandStateXORWOW *)                              
      0.0            2,336          1      2,336.0      2,336.0      2,336      2,336          0.0  void timemachine::k_compute_log_weights_from_energies<float>(int, T1, const __int128 *, T1 *)       
      0.0            2,079          1      2,079.0      2,079.0      2,079      2,079          0.0  void timemachine::k_compute_box_volume<float>(const double *, T1 *)                                 
      0.0            1,760          1      1,760.0      1,760.0      1,760      1,760          0.0  timemachine::k_arange(int, int *, int)                                                              

[6/7] Executing 'gpumemtimesum' stats report

CUDA Memory Operation Statistics (by time):

 Time (%)  Total Time (ns)  Count   Avg (ns)  Med (ns)  Min (ns)  Max (ns)   StdDev (ns)      Operation     
 --------  ---------------  ------  --------  --------  --------  ---------  -----------  ------------------
     31.6       55,441,222  28,634   1,936.2   1,952.0     1,472     13,088        340.0  [CUDA memcpy DtoD]
     27.0       47,421,508  38,034   1,246.8     864.0       511    185,792      3,210.6  [CUDA memcpy DtoH]
     22.0       38,632,289  18,849   2,049.6   2,208.0       735     81,312        887.1  [CUDA memset]     
     19.3       33,913,640     669  50,693.0  35,744.0       736  1,378,273    163,245.2  [CUDA memcpy HtoD]

[7/7] Executing 'gpumemsizesum' stats report

CUDA Memory Operation Statistics (by size):

 Total (MB)  Count   Avg (MB)  Med (MB)  Min (MB)  Max (MB)  StdDev (MB)      Operation     
 ----------  ------  --------  --------  --------  --------  -----------  ------------------
  3,147.496  28,634     0.110     0.009     0.000     0.213        0.102  [CUDA memcpy DtoD]
  2,478.263  18,849     0.131     0.171     0.000    28.865        0.243  [CUDA memset]     
    202.479     669     0.303     0.212     0.000     8.389        0.996  [CUDA memcpy HtoD]
     72.439  38,034     0.002     0.000     0.000     1.206        0.021  [CUDA memcpy DtoH]
```

### PR
```
CUDA API Statistics:

 Time (%)  Total Time (ns)  Num Calls  Avg (ns)   Med (ns)   Min (ns)   Max (ns)    StdDev (ns)            Name           
 --------  ---------------  ---------  ---------  ---------  --------  -----------  -----------  -------------------------
     48.7    1,575,151,965    411,939    3,823.8    3,545.0     2,915    1,268,397      2,684.8  cudaLaunchKernel         
     25.4      820,327,593     25,013   32,796.0    1,806.0     1,450      512,163     66,810.8  cudaEventSynchronize     
      7.1      228,595,522     53,663    4,259.8    3,526.0     2,241      492,149      8,515.4  cudaMemcpyAsync          
      6.8      220,584,485      2,018  109,308.5    7,480.5       121  174,145,828  3,876,526.8  cudaMalloc               
      2.9       92,296,377     18,827    4,902.3    4,007.0     2,119       51,719      4,272.9  cudaMemsetAsync          
      2.8       91,593,025      1,175   77,951.5   48,597.0     5,102    1,327,707    117,099.7  cudaMemcpy               
      2.5       80,999,532    113,346      714.6      576.0       427       17,652        616.9  cudaEventRecord          
      2.1       66,836,774     88,420      755.9      619.0       451       22,192        659.3  cudaStreamWaitEvent      
      1.3       41,176,225      2,038   20,204.2    5,772.5       270   11,013,383    246,261.4  cudaFree                 
      0.2        5,152,540        392   13,144.2    6,842.0     1,751      416,895     30,094.5  cudaStreamSynchronize    
      0.1        4,103,122          5  820,624.4  159,303.0    98,116    3,251,814  1,365,335.4  cudaDeviceSynchronize    
      0.1        4,023,985         11  365,816.8   10,805.0     6,987    1,008,019    494,524.7  cudaHostAlloc            
      0.1        2,677,096         11  243,372.4   20,497.0     5,904    1,117,458    361,395.9  cudaFreeHost             
      0.0          283,974         37    7,675.0    2,986.0     2,323      124,723     20,187.4  cudaStreamCreateWithFlags
      0.0          223,894         37    6,051.2    3,695.0     2,576       26,551      4,847.0  cudaStreamDestroy        
      0.0          208,561         22    9,480.0    5,705.5     2,450       21,660      7,040.4  cudaMemset               
…